requirements.txt
fastapi
uvicorn
langchain
langchain-groq
pymongo
python-dotenv
pydantic
python-multipart